backend/requirements.txt
Flask==3.0.3
flask_sqlalchemy==3.1.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30
Flask-SocketIO==5.3.6
Flask-JWT-Extended==4.6.0
google-cloud-firestore==2.16.0
gunicorn==22.0.0
flask-cors==4.0.1
